+++ nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/command/laborinsurance/employment/EmpInsurBusBurRatioCommand.java
@@ -1,5 +1,7 @@
 package nts.uk.ctx.pr.core.app.command.laborinsurance.employment;
 
+import java.math.BigDecimal;
+
 import lombok.Value;
 
 @Value
@@ -7,8 +9,8 @@ public class EmpInsurBusBurRatioCommand {
     
     private String hisId;
     private int empPreRateId;
-    private Long indBdRatio;
-    private Long empContrRatio;
+    private BigDecimal indBdRatio;
+    private BigDecimal empContrRatio;
     private int perFracClass;
     private int busiOwFracClass;
 ;
